fix(代码回滚): 继续回滚.

+++ Dianzhu.BLL/BLLBusinessImage.cs
@@ -23,12 +23,17 @@ namespace Dianzhu.BLL
        public void Delete(Guid biId)
        {
            BusinessImage bi = DALBusinessImage.GetOne(biId);
+            if (bi == null) { return; }
            string filePath = HttpContext.Current.Server.MapPath(SiteConfig.BusinessImagePath+"/original/")+bi.ImageName;
-           if (File.Exists(filePath))
-           {
-               File.Delete(filePath);
-           }
-           DALBusinessImage.Delete(bi);
+
+            if (filePath != null) {
+                if (File.Exists(filePath))
+                {
+                    File.Delete(filePath);
+                }
+                DALBusinessImage.Delete(bi);
+            }
+
        }
        public string  Save(Guid businessId,System.Web.HttpPostedFileBase imageFile, Dianzhu.Model.Enums.enum_ImageType imageType)
        {
